Extract logics to their respective module
